requirements.txt
pydantic
sanic
sanic-ext
sanic-testing
PyJWT
aiomysql
validators